requirements.txt
fastapi==0.70.0
nltk==3.6.5
pydantic==1.8.2
quickdraw==0.2.0
SpeechRecognition==3.8.1
stanford_openie==1.2.0
monkeylearn==3.6.0